+++ core/test.cpp
@@ -8,13 +8,13 @@ using EasingType = automation::EasingType;
 using EasingDirection = automation::EasingDirection;
 using Automation = automation::Automation;
 
-const double kEpsilon = 1e-9;
+const float kEpsilon = 1e-9f;
 
 TEST_CASE("single segment")
 {
     int numValues = 2;
-    double values[] = {1.0, 2.0};
-    double durations[] = {2.0};
+    float values[] = {1.0f, 2.0f};
+    float durations[] = {2.0f};
     EasingFunction easingFunctions[] = {
         { EasingType::Linear, EasingDirection::In, 0 },
     };
@@ -24,20 +24,20 @@ TEST_CASE("single segment")
         .durations = durations,
         .easingFunctions = easingFunctions
     };
-    REQUIRE_THAT(automation::evaluate(&automation, -1.0), Catch::Matchers::WithinRel(1.0, kEpsilon));
-    REQUIRE_THAT(automation::evaluate(&automation, 0.0), Catch::Matchers::WithinRel(1.0, kEpsilon));
-    REQUIRE_THAT(automation::evaluate(&automation, 0.5), Catch::Matchers::WithinRel(1.25, kEpsilon));
-    REQUIRE_THAT(automation::evaluate(&automation, 1.0), Catch::Matchers::WithinRel(1.5, kEpsilon));
-    REQUIRE_THAT(automation::evaluate(&automation, 2.0), Catch::Matchers::WithinRel(2.0, kEpsilon));
-    REQUIRE_THAT(automation::evaluate(&automation, 3.0), Catch::Matchers::WithinRel(2.0, kEpsilon));
+    REQUIRE_THAT(automation::evaluate(&automation, -1.0f), Catch::Matchers::WithinRel(1.0f, kEpsilon));
+    REQUIRE_THAT(automation::evaluate(&automation, 0.0f), Catch::Matchers::WithinRel(1.0f, kEpsilon));
+    REQUIRE_THAT(automation::evaluate(&automation, 0.5f), Catch::Matchers::WithinRel(1.25f, kEpsilon));
+    REQUIRE_THAT(automation::evaluate(&automation, 1.0f), Catch::Matchers::WithinRel(1.5f, kEpsilon));
+    REQUIRE_THAT(automation::evaluate(&automation, 2.0f), Catch::Matchers::WithinRel(2.0f, kEpsilon));
+    REQUIRE_THAT(automation::evaluate(&automation, 3.0f), Catch::Matchers::WithinRel(2.0f, kEpsilon));
 }
 
 
 TEST_CASE("multiple segments")
 {
     int numValues = 3;
-    double values[] = {1.0, 5.0, -3.0};
-    double durations[] = {2.0, 1.0};
+    float values[] = {1.0f, 5.0f, -3.0f};
+    float durations[] = {2.0f, 1.0f};
     EasingFunction easingFunctions[] = {
         { EasingType::Linear, EasingDirection::In, 0 },
         { EasingType::Linear, EasingDirection::In, 0 },
@@ -48,19 +48,19 @@ TEST_CASE("multiple segments")
         .durations = durations,
         .easingFunctions = easingFunctions
     };
-    REQUIRE_THAT(automation::evaluate(&automation, -1.0), Catch::Matchers::WithinRel(1.0, kEpsilon));
-    REQUIRE_THAT(automation::evaluate(&automation, 0.0), Catch::Matchers::WithinRel(1.0, kEpsilon));
-    REQUIRE_THAT(automation::evaluate(&automation, 1.0), Catch::Matchers::WithinRel(3.0, kEpsilon));
-    REQUIRE_THAT(automation::evaluate(&automation, 2.0), Catch::Matchers::WithinRel(5.0, kEpsilon));
-    REQUIRE_THAT(automation::evaluate(&automation, 3.0), Catch::Matchers::WithinRel(-3.0, kEpsilon));
-    REQUIRE_THAT(automation::evaluate(&automation, 4.0), Catch::Matchers::WithinRel(-3.0, kEpsilon));
+    REQUIRE_THAT(automation::evaluate(&automation, -1.0f), Catch::Matchers::WithinRel(1.0f, kEpsilon));
+    REQUIRE_THAT(automation::evaluate(&automation, 0.0f), Catch::Matchers::WithinRel(1.0f, kEpsilon));
+    REQUIRE_THAT(automation::evaluate(&automation, 1.0f), Catch::Matchers::WithinRel(3.0f, kEpsilon));
+    REQUIRE_THAT(automation::evaluate(&automation, 2.0f), Catch::Matchers::WithinRel(5.0f, kEpsilon));
+    REQUIRE_THAT(automation::evaluate(&automation, 3.0f), Catch::Matchers::WithinRel(-3.0f, kEpsilon));
+    REQUIRE_THAT(automation::evaluate(&automation, 4.0f), Catch::Matchers::WithinRel(-3.0f, kEpsilon));
 }
 
 TEST_CASE("NORMALIZATION")
 {
     int numValues = 3;
-    double values[] = {1.0, 2.0, 1.0};
-    double durations[] = {2.0, 3.0};
+    float values[] = {1.0f, 2.0f, 1.0f};
+    float durations[] = {2.0f, 3.0f};
     EasingFunction easingFunctions[] = {
         { EasingType::Linear, EasingDirection::In, 0 },
         { EasingType::Linear, EasingDirection::In, 0 },
@@ -72,8 +72,8 @@ TEST_CASE("NORMALIZATION")
         .easingFunctions = easingFunctions
     };
     normalizeDurations(&automation);
-    REQUIRE_THAT(automation.durations[0], Catch::Matchers::WithinRel(2.0 / 5.0, kEpsilon));
-    REQUIRE_THAT(automation.durations[1], Catch::Matchers::WithinRel(3.0 / 5.0, kEpsilon));
+    REQUIRE_THAT(automation.durations[0], Catch::Matchers::WithinRel(2.0f / 5.0f, kEpsilon));
+    REQUIRE_THAT(automation.durations[1], Catch::Matchers::WithinRel(3.0f / 5.0f, kEpsilon));
 }
 
 TEST_CASE("enumeration")
